Refactoring: class GeoDataPointmap::AddPointsCommand moved to independent files.
